remove 'ProtocolToMetaModel' from builder classes
