+++ .changeset/funny-items-smoke.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-graph': major
+---
+
+**BREAKING CHANGE:** We no longer expose `V1_ProtocolToMetaModelConnectionBuilder` and `V1_ProtocolToMetaModelEmbeddedDataBuilder`. Instead, we directly expose the builder helper `V1_buildConnection()` and `V1_buildEmbeddedData()`.
