+++ .changeset/pretty-bananas-relate.md
@@ -1,0 +1,25 @@
+---
+'@finos/legend-application': patch
+'@finos/legend-art': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-extension-dsl-persistence': patch
+'@finos/legend-extension-dsl-persistence-cloud': patch
+'@finos/legend-extension-dsl-text': patch
+'@finos/legend-extension-external-language-morphir': patch
+'@finos/legend-extension-external-store-service': patch
+'@finos/legend-extension-mapping-generation': patch
+'@finos/legend-graph': patch
+'@finos/legend-manual-tests': patch
+'@finos/legend-query': patch
+'@finos/legend-query-app': patch
+'@finos/legend-query-deployment': patch
+'@finos/legend-shared': patch
+'@finos/legend-studio': patch
+'@finos/legend-studio-app': patch
+'@finos/legend-studio-deployment': patch
+'@finos/legend-studio-extension-management-toolkit': patch
+'@finos/legend-studio-extension-query-builder': patch
+'@finos/legend-taxonomy': patch
+'@finos/legend-taxonomy-deployment': patch
+---
