+++ packages/legend-extension-dsl-persistence/src/models/protocols/pure/v1/transformation/pureGraph/to/V1_PersistenceBuilder.ts
@@ -159,9 +159,9 @@ import {
   type Database,
   type PackageableElementImplicitReference,
   type V1_GraphBuilderContext,
-  V1_ProtocolToMetaModelConnectionBuilder,
   V1_buildFullPath,
   optionalizePackageableElementReference,
+  V1_buildConnection,
 } from '@finos/legend-graph';
 import {
   guaranteeNonEmptyString,
@@ -206,9 +206,7 @@ export const V1_buildRelationalSink = (
     ) as PackageableElementImplicitReference<Database>;
   }
   if (protocol.connection) {
-    sink.connection = protocol.connection.accept_ConnectionVisitor(
-      new V1_ProtocolToMetaModelConnectionBuilder(context),
-    );
+    sink.connection = V1_buildConnection(protocol.connection, context);
   }
   return sink;
 };
@@ -223,9 +221,7 @@ export const V1_buildObjectStorageSink = (
     false,
   ) as PackageableElementImplicitReference<Binding>;
   if (protocol.connection) {
-    sink.connection = protocol.connection.accept_ConnectionVisitor(
-      new V1_ProtocolToMetaModelConnectionBuilder(context),
-    );
+    sink.connection = V1_buildConnection(protocol.connection, context);
   }
   return sink;
 };
