+++ packages/legend-graph/src/models/protocols/pure/v1/transformation/pureGraph/to/V1_RawValueSpecificationBuilder.ts
@@ -17,14 +17,14 @@
 import type { RawValueSpecification } from '../../../../../../metamodels/pure/rawValueSpecification/RawValueSpecification.js';
 import { RawLambda } from '../../../../../../metamodels/pure/rawValueSpecification/RawLambda.js';
 import { RawVariableExpression } from '../../../../../../metamodels/pure/rawValueSpecification/RawVariableExpression.js';
-import type { V1_GraphBuilderContext } from '../../../transformation/pureGraph/to/V1_GraphBuilderContext.js';
+import type { V1_GraphBuilderContext } from './V1_GraphBuilderContext.js';
 import type { V1_RawValueSpecificationVisitor } from '../../../model/rawValueSpecification/V1_RawValueSpecification.js';
 import type { V1_RawLambda } from '../../../model/rawValueSpecification/V1_RawLambda.js';
 import type { V1_RawVariable } from '../../../model/rawValueSpecification/V1_RawVariable.js';
 import type { V1_RawPrimitiveInstanceValue } from '../../../model/rawValueSpecification/V1_RawPrimitiveInstanceValue.js';
 import { RawPrimitiveInstanceValue } from '../../../../../../metamodels/pure/rawValueSpecification/RawPrimitiveInstanceValue.js';
 
-export class V1_ProtocolToMetaModelRawValueSpecificationBuilder
+export class V1_RawValueSpecificationBuilder
   implements V1_RawValueSpecificationVisitor<RawValueSpecification>
 {
   context: V1_GraphBuilderContext;
